Adjust the prior commit here (epee: abstract_tcp_server2: resolve CID 203919 (DC.WEAK_CRYPTO))
Adjustment requested by Jagerman, -
It reinitializes std::random_device on each pass through the function, and that can block, and on
top of that doesn't give randomization at all (you might as well just use the random device results
directly rather than using a mt19937 for one single random value generation).

This code needs to be fixed to initialize the RNG from the random device *once*, by moving the
19937+seed construction into a function and then, in the code in the commit, doing:

    static std::mt19937 rng = get_seeded_mt19937();

so that the seeding happens just *once* and then the random values are retrieved repeatedly from the
RNG.

It's also slightly wrong in how it pulls a random value using `rng() % 50`: that produces a
*slightly* biased value (because the range of produced values is not a multiple of 50).  It should
do this instead, to be correct:

    int ms = std::uniform_int_distribution<>{250, 299}(rng);

In practice the bias here isn't going to be significant, but there is also no downside to doing it
correctly here.