Fix fee logic
The fee logic here was not right: if given 0 for both the explicit and
default priorities it would calculate an unimportant priority.  Nothing
actually passes in 0 when doing a blink so this wasn't technically a
problem for sending transactions, but was still wrong.  (Fixing this
*without* the previous simplewallet fix would just end up using a blink
fee with a non-blink tx, so it is the more important fix).