Simplify is_local_address to just look for localhost
Doing a DNS resolution here seems like overkill, and even if you get a
resolution to a non-public IP that doesn't mean it's local/trusted (for
example, you could resolve a .loki address, get 10.0.0.3, and then
specify that).  So just hard-code common ways of specifying localhost
instead, which is really the main point of this.