Backport Fix bl.* commands from pulse to dev
bl.* weren't being registered for non-service nodes.  This fixes it by
moving endpoint setup from quorumnet_new (SN-only) to a new
quorumnet_init that does endpoint setup (for both SNs and non-SNs).

(This is a backport of the changes applied on the pulse branch to get
these fixes into `dev` for testing).