Make Monero happier with the libsodium-provided privkey
Passing the libsodium-derived private key into Monero was fine *except*
when you try to generate a pubkey from it via secret_key_to_public_key:
Monero for some reason has an extra check during pubkey generation that
requires the privkey is less than the basepoint.  (This check is
pointless because a value above the basepoint is perfectly acceptable:
all operations are always mod basepoint).

This is easy to work around by explicitly doing a mod L on the private
key value before setting it in the Monero privkey.  (This results in
exactly the same pubkey, and an exactly equivalent privkey).