Pulse: On signed block message, provide 2 signatures
- Previously we just submitted 1 signature that signed the contents of
the final block that required us to delay signature verification,
because, if we received the message before we were in the final stage we
would have to delay the verification because we have insufficient data
to verify the signature.

This means that when someone in the quorum receives and relays the
message, they can tamper the message and make it invalid (by changing
the round to something invalid for example) and cause other nodes in the
quorum to reject it, eventually, recording that the Service Node didn't
participate in the round and bias Service Nodes to decommissioning.

Instead of taking the shortcut and providing only 1 signature, we do the
same thing we do with all the other messages,

1. We signed the contents of the message- this proves that the message
originated from the Service Node it claims to have come from (preventing
any tampering).

2. The 2nd signature actually is the signature that signs the final
block and is included in the block for propagation in the network.

Doing so patches up the ability for intermediate relay nodes from tampering
the message.