Fix segfault in rpc shutdown
Fixes a couple of issues that can cause a segfault:

- we can get a shutdown callback double-scheduled, which would lead to a
segfault because we'd try freeing listening socket pointers that have
already been freed.  Clear the variable properly so that if we get
called more than once we don't actually do anything.

- Don't store nullptr failed listen sockets. When uWebSockets fails to
bind it gives us a nullptr; we were storing this for some reason, but
shouldn't be (because later we try to dereference/free it).  With the
recent change this started triggering much more frequently because the
build CIs don't have IPv6 (and so the optional bind to [::1]:22023 would
fail).